# Conflicts:
#	script/deployment/deployment-addresses.json
#	script/shell/deploy.sh

+++ src/interfaces/legacy/Treasury.sol
@@ -4,33 +4,33 @@ pragma solidity >=0.7.5;
 interface ILegacyTreasury {
     function bondCalculator(address _address) external view returns (address);
 
-    function deposit(uint256 _amount, address _token, uint256 _profit) external returns (uint256);
+    function deposit(uint _amount, address _token, uint _profit) external returns (uint);
 
-    function withdraw(uint256 _amount, address _token) external;
+    function withdraw(uint _amount, address _token) external;
 
-    function depositERC721(address _token, uint256 _tokenId) external;
+    function depositERC721(address _token, uint _tokenId) external;
 
-    function withdrawERC721(address _token, uint256 _tokenId) external;
+    function withdrawERC721(address _token, uint _tokenId) external;
 
-    function tokenValue(address _token, uint256 _amount) external view returns (uint256 value_);
+    function tokenValue(address _token, uint _amount) external view returns (uint value_);
 
-    function mint(address _recipient, uint256 _amount) external;
+    function mint(address _recipient, uint _amount) external;
 
-    function manage(address _token, uint256 _amount) external;
+    function manage(address _token, uint _amount) external;
 
-    function allocatorManage(address _token, uint256 _amount) external;
+    function allocatorManage(address _token, uint _amount) external;
 
-    function claimNFTXRewards(address _liquidityStaking, uint256 _vaultId, address _rewardToken) external;
+    function claimNFTXRewards(address _liquidityStaking, uint _vaultId, address _rewardToken) external;
 
-    function incurDebt(uint256 amount_, address token_) external;
+    function incurDebt(uint amount_, address token_) external;
 
-    function repayDebtWithReserve(uint256 amount_, address token_) external;
+    function repayDebtWithReserve(uint amount_, address token_) external;
 
-    function excessReserves() external view returns (uint256);
+    function excessReserves() external view returns (uint);
 
-    function riskOffValuation(address _token) external view returns (uint256);
+    function riskOffValuation(address _token) external view returns (uint);
 
-    function baseSupply() external view returns (uint256);
+    function baseSupply() external view returns (uint);
 
     function enable(uint8 _status, address _address, address _calculator) external;
 }
