+++ src/interfaces/uniswap/IUniversalRouter.sol
@@ -3,11 +3,12 @@ pragma solidity ^0.8.17;
 
 import {IERC721Receiver} from '@openzeppelin/contracts/token/ERC721/IERC721Receiver.sol';
 import {IERC1155Receiver} from '@openzeppelin/contracts/token/ERC1155/IERC1155Receiver.sol';
-import {IRewardsCollector} from './IRewardsCollector.sol';
+
+import {IRewardsCollector} from '@floor-interfaces/uniswap/IRewardsCollector.sol';
 
 interface IUniversalRouter is IRewardsCollector, IERC721Receiver, IERC1155Receiver {
     /// @notice Thrown when a required command has failed
-    error ExecutionFailed(uint256 commandIndex, bytes message);
+    error ExecutionFailed(uint commandIndex, bytes message);
 
     /// @notice Thrown when attempting to send ETH directly to the contract
     error ETHNotAccepted();
@@ -22,5 +23,5 @@ interface IUniversalRouter is IRewardsCollector, IERC721Receiver, IERC1155Receiv
     /// @param commands A set of concatenated commands, each 1 byte in length
     /// @param inputs An array of byte strings containing abi encoded inputs for each command
     /// @param deadline The deadline by which the transaction must be executed
-    function execute(bytes calldata commands, bytes[] calldata inputs, uint256 deadline) external payable;
+    function execute(bytes calldata commands, bytes[] calldata inputs, uint deadline) external payable;
 }
